Added a dictionary which stores the aspect connection names as the keys and the number of times that connection is traversed as for the value.
A get function is implemented to return the current dictionary when called.

An incomplete, but functioning method to compute this traversal tracker is implemented. This method is called "generate_connection_traversal_count". This will need to be adjusted and refined to account for factors in data. It has yet to be fully determined what factors this includes.